Fixed undefined url to item detail page #1327
